Update to latest changes from upstream, fix correctness on MS Visual Studio

+++ crypto_kem/frodokem1344aes/clean/noise.c
@@ -20,9 +20,9 @@ void PQCLEAN_FRODOKEM1344AES_CLEAN_sample_n(uint16_t *s, size_t n) {
     unsigned int j;
 
     for (i = 0; i < n; ++i) {
-        uint8_t sample = 0;
+        uint16_t sample = 0;
         uint16_t prnd = s[i] >> 1;    // Drop the least significant bit
-        uint8_t sign = s[i] & 0x1;    // Pick the least significant bit
+        uint16_t sign = s[i] & 0x1;    // Pick the least significant bit
 
         // No need to compare with the last value.
         for (j = 0; j < (unsigned int)(CDF_TABLE_LEN - 1); j++) {
